pyproject.toml
[project]
name = "flask-music-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "blinker>=1.8.2",
    "click>=8.2.2",
    "flask>=3.0.3",
    "gunicorn>=23.0.0",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.4",
    "markupsafe>=3.0.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "ruamel-yaml>=0.18.14",
    "werkzeug>=3.1.3",
]
